+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Release date `UNRELEASED`
 
-
+- (docs api exchange): unify arglists, add link to javadocs
 
 ## `1.4.0-125`